MS-29 feat(registration-form): redirect users to the application's main page upon successful account creation and automatic login
