+++ src/api/api.ts
@@ -1,5 +1,14 @@
-import ctpClient from "BuildClient";
 import { CustomerDraft, createApiBuilderFromCtpClient } from "@commercetools/platform-sdk";
+import generateAnonymousSessionFlow from "./anonymousClientBuilder";
+
+let ctpClient;
+if (!localStorage.getItem("token")) {
+  ctpClient = generateAnonymousSessionFlow();
+  console.log("аноним");
+}
+if (localStorage.getItem("token")) {
+  ctpClient = console.log("рефреш"); // generateAnonymousSessionFlow();
+}
 
 const apiRoot = createApiBuilderFromCtpClient(ctpClient).withProjectKey({ projectKey: "steps-moon-store" });
